refactor: isolate config handling
