+++ ninox/image_description.py
@@ -4,7 +4,6 @@ import base64
 import json
 import mimetypes
 import os
-import tomllib
 from pathlib import Path
 
 import click
@@ -12,61 +11,8 @@ from openai import OpenAI
 from openai.types.responses import ResponseInputImageParam, ResponseInputTextParam
 from openai.types.responses.response_input_param import Message
 from PIL import Image
-from pydantic import BaseModel
 
-
-class OpenAITokens(BaseModel):
-    """
-    Holds OpenAI tokens from config.
-    AI: Generated by ChatGPT
-    """
-
-    open: str
-    closed: str
-
-
-class TokensConfig(BaseModel):
-    """
-    Wrapper for token groups.
-
-    AI: Generated by ChatGPT
-    """
-
-    openai: OpenAITokens
-
-
-class Config(BaseModel):
-    """
-    Root configuration model.
-
-    AI: Generated by ChatGPT
-    """
-
-    tokens: TokensConfig
-
-
-def load_config(config_path: Path | str) -> Config:
-    """
-    Load and parse a TOML config file with pydantic.
-
-    Args:
-        config_path: Path to the TOML config file.
-
-    Returns:
-        Config instance populated from TOML data.
-
-    Raises:
-        FileNotFoundError: If the file does not exist.
-        tomllib.TOMLDecodeError: If TOML is invalid.
-
-    AI: Generated by ChatGPT
-    """
-    path = Path(config_path).expanduser()
-    if not path.is_file():
-        raise FileNotFoundError(f"Config file not found: {path}")
-    with path.open("rb") as f:
-        data = tomllib.load(f)
-    return Config(**data)
+from .config import load_config
 
 
 def get_image_description(
